refactor(knowledge): rename graph.py → task_graph.py, semantic_graph.py → ast_symbol_graph.py (audit-177)
The three files under core/knowledge/ that all end in "graph" describe
three different domains. Rename the two ambiguous ones to match their
purpose so readers don't have to open them to find out:

- graph.py → task_graph.py (task-dependency DAG, in-memory, rebuilt per
  tick). 7 prod importers updated.
- semantic_graph.py → ast_symbol_graph.py (AST-derived symbol graph,
  in-memory; seeds knowledge_graph). 1 prod importer updated.
- knowledge_graph.py stays (SQLite codebase symbol store). 2 prod
  importers updated to the explicit sub-package path.

_REDIRECT_MAP keeps old names ("graph", "semantic_graph") as shims
pointing at the new targets so any external caller still using
`bernstein.core.graph` / `bernstein.core.semantic_graph` continues to
work. Added primary keys "task_graph" and "ast_symbol_graph" for the
new flat names. (_REDIRECT_MAP edits already landed via audit-169.)

Added docs/architecture/graphs.md documenting the three graph domains,
their storage, and consumers.
