+++ src/bernstein/cli/commands/graph_cmd.py
@@ -9,7 +9,7 @@ import click
 import httpx
 
 from bernstein.cli.helpers import SERVER_URL, auth_headers, console
-from bernstein.core.knowledge_graph import query_impact
+from bernstein.core.knowledge.knowledge_graph import query_impact
 
 # Shared cast-type constants to avoid string duplication (Sonar S1192).
 _CAST_LIST_DICT_STR_ANY = "list[dict[str, Any]]"
