+++ src/bernstein/core/orchestration/orchestrator.py
@@ -69,7 +69,7 @@ from bernstein.core.fast_path import (
     load_fast_path_config,
 )
 from bernstein.core.file_locks import FileLockManager
-from bernstein.core.graph import TaskGraph
+from bernstein.core.knowledge.task_graph import TaskGraph
 from bernstein.core.hook_events import HookEvent
 from bernstein.core.incident import IncidentManager
 from bernstein.core.manifest import build_manifest, save_manifest
