+++ tests/unit/test_ast_symbol_graph.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from pathlib import Path
 
-import bernstein.core.semantic_graph as semantic_graph
+import bernstein.core.knowledge.ast_symbol_graph as semantic_graph
 import pytest
-from bernstein.core.semantic_graph import (
+from bernstein.core.knowledge.ast_symbol_graph import (
     SemanticGraph,
     SymbolEdge,
     SymbolNode,
