+++ tests/unit/test_knowledge_graph.py
@@ -5,12 +5,16 @@ from __future__ import annotations
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-import bernstein.core.knowledge_graph as knowledge_graph
-import bernstein.core.semantic_graph as semantic_graph
+import bernstein.core.knowledge.ast_symbol_graph as semantic_graph
+import bernstein.core.knowledge.knowledge_graph as knowledge_graph
 import pytest
 import pytest_asyncio
 from bernstein.cli.graph_cmd import graph_group
-from bernstein.core.knowledge_graph import build_knowledge_graph, get_or_build_knowledge_graph, query_impact
+from bernstein.core.knowledge.knowledge_graph import (
+    build_knowledge_graph,
+    get_or_build_knowledge_graph,
+    query_impact,
+)
 from click.testing import CliRunner
 from httpx import ASGITransport, AsyncClient
