+++ tests/unit/test_task_lifecycle.py
@@ -12,7 +12,7 @@ from unittest.mock import MagicMock, patch
 
 import httpx
 from bernstein.core.convergence_guard import ConvergenceGuard
-from bernstein.core.graph import TaskGraph
+from bernstein.core.knowledge.task_graph import TaskGraph
 from bernstein.core.models import (
     AgentSession,
     CompletionSignal,
